Fixes TeX module for Emacs
Adds support for changing the org and tex config file path. Also makes
extra packages needed by org and tex customizable outside of the
module definition.

Additionally this makes the launchd service for Emacs on darwin use
the finalPackge output instead of package. This improves compatibility
with existing home-manager modules as well as my additions for
controlling whether or not org and tex are configured.